requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.0
aiogram==3.2.0
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.19.0
psycopg2-binary==2.9.9
websockets==12.0
Pillow==10.1.0
pytz==2024.1